// src/typechain/out/UniswapV4Adapter.ts
export type PathKeyStruct = {
  intermediateCurrency: AddressLike;
  fee: BigNumberish;
  tickSpacing: BigNumberish;
  hooks: AddressLike;
  hookData: BytesLike;
};

export type PathKeyStructOutput = [
  intermediateCurrency: string,
  fee: bigint,
  tickSpacing: bigint,
  hooks: string,
  hookData: string
] & {
  intermediateCurrency: string;
  fee: bigint;
  tickSpacing: bigint;
  hooks: string;
  hookData: string;
};

export type SwapV4Struct = {
  amountToSend: BigNumberish;
  amountToReceive: BigNumberish;
  currencyIn: AddressLike;
  path: PathKeyStruct[];
  deadline: BigNumberish;
};